Small Physics Fine-tuning. Backup for experimental changes

+++ Simulation/src/brunner/jens/main/DrawingComponent.java
@@ -43,9 +43,9 @@ public class DrawingComponent extends JComponent
 	}
 	
 	public static void paintBody(Vector2 position, Vector2 vel) {
-		float radius = Main.scaleFactor*4/2;
+		double radius = Main.scaleFactor*4/2;
 		if(radius < 1) radius = 1;
-		g2.fill(new Ellipse2D.Float(position.x-radius, position.y-radius, radius*2, radius*2));
+		g2.fill(new Ellipse2D.Double(position.x-radius, position.y-radius, radius*2, radius*2));
 		if(Main.showVelocityArrows)
 		{
 		//To draw the velocity vectors, we create a line.
